pyproject.toml
[tool.poetry]
name = "Pipeboard"
version = "1.0.0"
description = "Integração Pipedrive+Metabase via datalake Postgres com streaming ETL com Prefect."
authors = ["pavcob"]
readme = ["README.md"]

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
prefect = {extras = ["docker"], version = "^3.4.1"}
requests = "2.32.3"
python-decouple = "3.8"
python-dotenv = "1.1.0"
pydantic = '2.11.4'
prometheus-client = "0.21.1"
psycopg2  = '2.9.10'
httpx='0.28.1'
tenacity='9.1.2'
python-json-logger='3.3.0'
prefect-sqlalchemy = "0.5.2"
pandas = "2.2.3"
numpy = "2.2.5"
structlog = "25.3.0"
geopy = "2.4.1"
pybreaker = "1.3.0"
psutil="7.0.0"
prefect-docker = "0.6.4"
redis = "^6.1.0"
pyarrow = "^20.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.5"
pytest-django = "4.11.1"      
ruff = "0.11.10"
mypy = "1.15.0"
types-requests = "^2.32"  

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
